% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_lasso.R
\name{extreme_lasso}
\alias{extreme_lasso}
\title{Estimation algorithm of the Extreme Lasso problem.}
\usage{
extreme_lasso(
  resp,
  pred_mat,
  sub_power,
  abstol = 0.001,
  reltol = 0.001,
  init_guess = NULL,
  init_step = 1,
  shrink = 0.5,
  lambda = 1,
  maxit = 100
)
}
\arguments{
\item{resp}{n by 1 response matrix.}

\item{pred_mat}{n by p predictor matrix.}

\item{sub_power}{even integer, power of Subbotin distribution.}

\item{abstol}{maximum absolute error.}

\item{reltol}{maximum relative error.}

\item{init_guess}{initialization of parameter estimates.}

\item{init_step}{initial step size scale parameter}

\item{shrink}{backtracking shrinkage parameter}

\item{lambda}{L1 regularization penalty.}

\item{maxit}{maximum iterations of gradient descent.}
}
\value{
A list with the following elements:
\itemize{
\item pars: p by 1 matrix, final parameter estimates
\item sub_power: power of Subbotin distribution.
\item iterations: number of iterations run.
}
}
\description{
Estimates for the Extreme Lasso problem; used for neighborhood selection. Uses proximal gradient descent with backtracking.
}
